+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Suggest replying from the home server when guest clicks on a reply button.
 - Added icon to "Subscribe" button on profile page.
 - Added setting for shortening time unit in post timestamp.
+- Added session manager.
 
 ### Changed